# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Egi Ramadhan FD Test API
  description: Full-stack application API with authentication and book management
  version: 1.0.0
  contact:
    name: Egi Ramadhan
    email: egi@example.com
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /api/books:
    get:
      tags:
        - Books
      summary: Get books list
      parameters:
        - name: author
          in: query
          schema:
            type: string
        - name: rating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Books list retrieved successfully
    post:
      tags:
        - Books
      summary: Create a new book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - author
                - description
                - rating
              properties:
                title:
                  type: string
                author:
                  type: string
                description:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                thumbnail:
                  type: string
                  format: binary
                thumbnailUrl:
                  type: string
                  format: uri
      responses:
        '201':
          description: Book created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/books/{id}:
    get:
      tags:
        - Books
      summary: Get book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book retrieved successfully
        '404':
          description: Book not found
    patch:
      tags:
        - Books
      summary: Update book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                description:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                thumbnail:
                  type: string
                  format: binary
                thumbnailUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: Book updated successfully
        '404':
          description: Book not found
        '401':
          description: Unauthorized
        '403':
          description: Not authorized to update this book
    delete:
      tags:
        - Books
      summary: Delete book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book deleted successfully
        '404':
          description: Book not found
        '401':
          description: Unauthorized
        '403':
          description: Not authorized to delete this book
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...